backend/requirements.txt
Flask==2.3.2
flask-cors==3.0.10
firebase-admin==6.1.0
google-generativeai==0.2.0
termcolor==2.3.0
